+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - redefine authorizeEventServiceOfferAction.result as AggregateOffer
 - optimize action.ISearchConditions
 - make authorizeEventServiceOfferAction.object.acceptedOffer optional
+- optimize returnPayTransaction task
 
 ### Deprecated